[virtdisk-ss] Extract utils to separate files
