+++ src/platform/backends/hyperv_api/CMakeLists.txt
@@ -64,6 +64,8 @@ if(WIN32)
     virtdisk/virtdisk_snapshot.cpp
     virtdisk/virtdisk_api.cpp
     virtdisk/virtdisk_guid.cpp
+    virtdisk/virtdisk_child_rebuild.cpp
+    virtdisk/virtdisk_utils.cpp
     hcs_virtual_machine.cpp
     hcs_virtual_machine_factory.cpp
   )
