+++ src/platform/backends/hyperv_api/virtdisk/virtdisk_exceptions.h
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) Canonical, Ltd.
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; version 3.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#pragma once
+
+#include <multipass/exceptions/formatted_exception_base.h>
+
+namespace multipass::hyperv::virtdisk
+{
+struct CreateVirtdiskSnapshotError : FormattedExceptionBase<std::system_error>
+{
+    using FormattedExceptionBase::FormattedExceptionBase;
+};
+
+} // namespace multipass::hyperv::virtdisk
